v2/requirements.classify.txt
requests==2.28.1
google-api-core==2.11.0
google-auth==2.15.0
google-auth-oauthlib==0.4.6
google-cloud-storage==2.6.0
pytz==2022.6
pillow==9.3.0
google-api-python-client==2.68.0
tensorflow==2.10.1
flask==2.2.2
tweepy==4.12.1
astral==3.2

